+++ lib/lightstep/propagation.rb
@@ -6,15 +6,14 @@ require 'lightstep/propagation/b3_propagator'
 module LightStep
   module Propagation
     PROPAGATOR_MAP = {
-      lightstep: LightStepPropagator,
-      b3: B3Propagator
+      lightstep: LightStepPropagator
     }
 
     class << self
       # Constructs a propagator instance from the given propagator name. If the
       # name is unknown returns the LightStepPropagator as a default
       #
-      # @param [Symbol, String] propagator_name One of :lightstep or :b3
+      # @param [Symbol, String] propagator_name
       # @return [Propagator]
       def [](propagator_name)
         klass = PROPAGATOR_MAP[propagator_name.to_sym] || LightStepPropagator