+++ CHANGELOG.md
@@ -3,13 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## v0.16.1
-### Bugfix
-- The tracer now closes the active scope or finishes the active span even if an error is raised from the yielded code.
 
 ## v0.16.0
 ### Added
 - The tracer now supports B3 context propagation. Propagation can be set by using the `propagator` keyword argument to `LightStep.configure`. Valid values are `:lightstep` (default), and `:b3`.
+- The tracer now closes the active scope or finishes the active span even if an error is raised from the yielded code.
 
 ## v0.15.0
 ### Added