+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.16.1
+### Changed
+- Bump json from 2.2.0 to 2.5.1 ([#98](https://github.com/lightstep/lightstep-tracer-ruby/pull/98))
+- Bump rack from 2.0.7 to 2.2.3 ([#95](https://github.com/lightstep/lightstep-tracer-ruby/pull/98))
+- Update rake requirement from ~> 11.3 to ~> 13.0 ([#94](https://github.com/lightstep/lightstep-tracer-ruby/pull/98))
+
 ## v0.16.0
 ### Added
 - The tracer now supports B3 context propagation. Propagation can be set by using the `propagator` keyword argument to `LightStep.configure`. Valid values are `:lightstep` (default), and `:b3`.